Refs #31349. Update FDA to use the Basic fitting widget.
