+++ scripts/Muon/GUI/Common/fitting_widgets/basic_fitting/basic_fitting_model.py
@@ -4,12 +4,14 @@
 #   NScD Oak Ridge National Laboratory, European Spallation Source,
 #   Institut Laue - Langevin & CSNS, Institute of High Energy Physics, CAS
 # SPDX - License - Identifier: GPL - 3.0 +
-from mantid import AlgorithmManager, logger
+from mantid import AlgorithmManager, AnalysisDataService, logger
 from mantid.api import CompositeFunction, IAlgorithm, IFunction
 from mantid.simpleapi import CopyLogs, EvaluateFunction, RenameWorkspace
 
 from Muon.GUI.Common.ADSHandler.ADS_calls import check_if_workspace_exist, retrieve_ws
-from Muon.GUI.Common.ADSHandler.workspace_naming import create_fitted_workspace_name, create_parameter_table_name
+from Muon.GUI.Common.ADSHandler.workspace_naming import (create_fitted_workspace_name, create_parameter_table_name,
+                                                         get_group_or_pair_from_name,
+                                                         get_run_number_from_workspace_name)
 from Muon.GUI.Common.ADSHandler.muon_workspace_wrapper import MuonWorkspaceWrapper
 from Muon.GUI.Common.contexts.fitting_context import FitInformation
 from Muon.GUI.Common.contexts.muon_context import MuonContext
@@ -58,6 +60,9 @@ class BasicFittingModel:
     def __init__(self, context: MuonContext, is_frequency_domain: bool = False):
         """Initializes the model with empty fit data."""
         self.context = context
+
+        self._x_data_type = self.context._frequency_context.plot_type if is_frequency_domain else "None"
+        self._group_or_pair_index = {}
 
         self._default_end_x = DEFAULT_FDA_END_X if is_frequency_domain else DEFAULT_MA_END_X
 
@@ -395,11 +400,6 @@ class BasicFittingModel:
         """Returns true if the fitting mode is simultaneous. Override this method if you require simultaneous."""
         return False
 
-    @property
-    def global_parameters(self) -> list:
-        """Returns the global parameters stored in the model. Override this method if you require global parameters."""
-        return []
-
     def update_parameter_value(self, full_parameter: str, value: float) -> None:
         """Update the value of a parameter in the fit function."""
         if self.current_single_fit_function is not None:
@@ -542,9 +542,59 @@ class BasicFittingModel:
         else:
             return []
 
-    def get_workspace_names_to_display_from_context(self) -> None:
-        """Returns the workspace names to display in the view and store in the model."""
-        raise NotImplementedError("This method must be overridden by a child class.")
+    def get_workspace_names_to_display_from_context(self) -> list:
+        """Returns the workspace names to display in the view based on the selected run and group/pair options."""
+        runs, groups_and_pairs = self.get_selected_runs_groups_and_pairs()
+
+        display_workspaces = []
+        for group_and_pair in groups_and_pairs:
+            display_workspaces += self._get_workspace_names_to_display_from_context(runs, group_and_pair)
+
+        return self._sort_workspace_names(display_workspaces)
+
+    def _get_workspace_names_to_display_from_context(self, runs: list, group_and_pair: str) -> list:
+        """Returns the workspace names for the given runs and group/pair to be displayed in the view."""
+        return self.context.get_names_of_workspaces_to_fit(runs=runs, group_and_pair=group_and_pair,
+                                                           rebin=not self.fit_to_raw, freq=self._x_data_type)
+
+    def _sort_workspace_names(self, workspace_names: list) -> list:
+        """Sort the workspace names and check the workspaces exist in the ADS."""
+        workspace_names = list(set(self._check_data_exists(workspace_names)))
+        if len(workspace_names) > 1:
+            workspace_names.sort(key=self._workspace_list_sorter)
+        return workspace_names
+
+    def _workspace_list_sorter(self, workspace_name: str) -> int:
+        """Used to sort a list of workspace names based on run number and group/pair name."""
+        run_number = get_run_number_from_workspace_name(workspace_name, self.context.data_context.instrument)
+        grp_pair_number = self._transform_group_or_pair_to_float(workspace_name)
+        return int(run_number) + grp_pair_number
+
+    def _transform_group_or_pair_to_float(self, workspace_name: str) -> int:
+        """Converts the workspace group or pair name to a float which is used in sorting the workspace list."""
+        group_or_pair_name = get_group_or_pair_from_name(workspace_name)
+        if group_or_pair_name not in self._group_or_pair_index:
+            self._group_or_pair_index[group_or_pair_name] = len(self._group_or_pair_index)
+
+        group_or_pair_values = list(self._group_or_pair_index.values())
+        if len(self._group_or_pair_index) > 1:
+            return ((self._group_or_pair_index[group_or_pair_name] - group_or_pair_values[0])
+                    / (group_or_pair_values[-1] - group_or_pair_values[0])) * 0.99
+        else:
+            return 0
+
+    @staticmethod
+    def _check_data_exists(workspace_names: list) -> list:
+        """Returns only the workspace names that exist in the ADS."""
+        return [workspace_name for workspace_name in workspace_names if AnalysisDataService.doesExist(workspace_name)]
+
+    def get_selected_runs_groups_and_pairs(self) -> tuple:
+        """Returns the runs, groups and pairs to use for single fit mode."""
+        return "All", self._get_selected_groups_and_pairs()
+
+    def _get_selected_groups_and_pairs(self) -> list:
+        """Returns the groups and pairs currently selected in the context."""
+        return self.context.group_pair_context.selected_groups_and_pairs
 
     def perform_fit(self) -> tuple:
         """Performs a single fit and returns the resulting function, status and chi squared."""
