+++ scripts/Muon/GUI/Common/fitting_widgets/general_fitting/general_fitting_model.py
@@ -9,8 +9,7 @@ from mantid.simpleapi import RenameWorkspace, CopyLogs
 
 from Muon.GUI.Common.ADSHandler.workspace_naming import (create_fitted_workspace_name,
                                                          create_multi_domain_fitted_workspace_name,
-                                                         create_parameter_table_name, get_group_or_pair_from_name,
-                                                         get_run_number_from_workspace_name,
+                                                         create_parameter_table_name,
                                                          get_run_numbers_as_string_from_workspace_name)
 from Muon.GUI.Common.contexts.muon_context import MuonContext
 from Muon.GUI.Common.fitting_widgets.basic_fitting.basic_fitting_model import BasicFittingModel
@@ -25,10 +24,6 @@ class GeneralFittingModel(BasicFittingModel):
     def __init__(self, context: MuonContext, is_frequency_domain: bool = False):
         """Initialize the GeneralFittingModel with emtpy fit data."""
         super(GeneralFittingModel, self).__init__(context, is_frequency_domain)
-
-        self._x_data_type = self.context._frequency_context.plot_type if is_frequency_domain else "None"
-
-        self._group_or_pair_index = {}
 
         # This is a MultiDomainFunction if there are multiple domains in the function browser.
         self._simultaneous_fit_function = None
@@ -218,16 +213,6 @@ class GeneralFittingModel(BasicFittingModel):
             return self._get_selected_groups_and_pairs()
         return []
 
-    def get_workspace_names_to_display_from_context(self) -> list:
-        """Returns the workspace names to display in the view based on the selected run and group/pair options."""
-        runs, groups_and_pairs = self.get_selected_runs_groups_and_pairs()
-
-        display_workspaces = []
-        for group_and_pair in groups_and_pairs:
-            display_workspaces += self._get_workspace_names_to_display_from_context(runs, group_and_pair)
-
-        return self._sort_workspace_names(display_workspaces)
-
     def get_fit_function_parameters(self) -> list:
         """Returns the names of the fit parameters in the fit functions."""
         if self.simultaneous_fitting_mode:
@@ -264,7 +249,7 @@ class GeneralFittingModel(BasicFittingModel):
         if self.simultaneous_fitting_mode:
             return self._get_selected_runs_groups_and_pairs_for_simultaneous_fit_mode()
         else:
-            return self._get_selected_runs_groups_and_pairs_for_single_fit_mode()
+            return super().get_selected_runs_groups_and_pairs()
 
     def _get_selected_runs_groups_and_pairs_for_simultaneous_fit_mode(self) -> tuple:
         """Returns the runs, groups and pairs that are currently selected for simultaneous fit mode."""
@@ -275,14 +260,6 @@ class GeneralFittingModel(BasicFittingModel):
         elif self.simultaneous_fit_by == "Group/Pair":
             groups_and_pairs = [self.simultaneous_fit_by_specifier]
         return runs, groups_and_pairs
-
-    def _get_selected_runs_groups_and_pairs_for_single_fit_mode(self) -> tuple:
-        """Returns the runs, groups and pairs to use for single fit mode."""
-        return "All", self._get_selected_groups_and_pairs()
-
-    def _get_selected_groups_and_pairs(self) -> list:
-        """Returns the groups and pairs currently selected in the context."""
-        return self.context.group_pair_context.selected_groups_and_pairs
 
     def _get_selected_runs(self) -> list:
         """Returns an ordered list of run numbers currently selected in the context."""
@@ -304,42 +281,6 @@ class GeneralFittingModel(BasicFittingModel):
         workspace_list = self.context.data_context.current_data["OutputWorkspace"]
         return [get_run_numbers_as_string_from_workspace_name(workspace.workspace_name, instrument)
                 for workspace in workspace_list]
-
-    def _get_workspace_names_to_display_from_context(self, runs: list, group_and_pair: str) -> list:
-        """Returns the workspace names for the given runs and group/pair to be displayed in the view."""
-        return self.context.get_names_of_workspaces_to_fit(runs=runs, group_and_pair=group_and_pair,
-                                                           rebin=not self.fit_to_raw, freq=self._x_data_type)
-
-    def _sort_workspace_names(self, workspace_names: list) -> list:
-        """Sort the workspace names and check the workspaces exist in the ADS."""
-        workspace_names = list(set(self._check_data_exists(workspace_names)))
-        if len(workspace_names) > 1:
-            workspace_names.sort(key=self._workspace_list_sorter)
-        return workspace_names
-
-    def _workspace_list_sorter(self, workspace_name: str) -> int:
-        """Used to sort a list of workspace names based on run number and group/pair name."""
-        run_number = get_run_number_from_workspace_name(workspace_name, self.context.data_context.instrument)
-        grp_pair_number = self._transform_group_or_pair_to_float(workspace_name)
-        return int(run_number) + grp_pair_number
-
-    def _transform_group_or_pair_to_float(self, workspace_name: str) -> int:
-        """Converts the workspace group or pair name to a float which is used in sorting the workspace list."""
-        group_or_pair_name = get_group_or_pair_from_name(workspace_name)
-        if group_or_pair_name not in self._group_or_pair_index:
-            self._group_or_pair_index[group_or_pair_name] = len(self._group_or_pair_index)
-
-        group_or_pair_values = list(self._group_or_pair_index.values())
-        if len(self._group_or_pair_index) > 1:
-            return ((self._group_or_pair_index[group_or_pair_name] - group_or_pair_values[0])
-                    / (group_or_pair_values[-1] - group_or_pair_values[0])) * 0.99
-        else:
-            return 0
-
-    @staticmethod
-    def _check_data_exists(workspace_names: list) -> list:
-        """Returns only the workspace names that exist in the ADS."""
-        return [workspace_name for workspace_name in workspace_names if AnalysisDataService.doesExist(workspace_name)]
 
     def perform_fit(self) -> tuple:
         """Performs a single or simultaneous fit and returns the resulting function, status and chi squared."""
